CBES-83 Plugin: VersionConflictEngineException in storeUUID
Motivation
==========
The intent of storeUUID is to create the document if possible, otherwise
fail silently. Instead what's happening is that if the UUID document
already exists, an exception is thrown and propagated all the way
to the Jetty ServletHandler, causing Couchbase to remove the
XDCR replication.

Modifications
=============
Catch and log any exception thrown by the document creation operation.

Result
======
Concurrent calls to storeUUID no longer cause the method to throw
an exception that removes the XDCR replication.