Broadcast RPC requests in parallel instead of sequentially
Motivation
==========
The `stopStreaming` request takes about two seconds. Broadcasting
it to all nodes was taking 2 seconds *per node*. In a timeout
situation, the broadcast duration could be the timeout duration
times the number of nodes.

Modifications
=============
When broadcasting a request to multiple nodes, schedule the requests
on separate threads.

Incidentally, unify the various broadcast methods so there's less
code duplication.

Change-Id: I7c2e5d4d30f05516265a44aeb703f43e932f9094
Reviewed-on: http://review.couchbase.org/104169
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>