Consul: Integration test infrastructure
Consolidate TestConnector and ConnectorTask into AsyncTask,
which is also used to run the connector in Consul mode.

Patch the connector config in the String domain instead of the POJO
domain. Need to do this in order to upload the patched config to Consul.

Add basic streaming test for Consul mode.

Change-Id: I43fa48bb8a3efedfd41cdce7686527bace9eaad5
Reviewed-on: http://review.couchbase.org/105557
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>