CBES-117 Connector exits on values that fail to parse
Motivation
==========
If a document contains an integer value larger than 2^63-1 an
IllegalArgumentException is thrown from RequestFactory.newIndexRequest.
The error happens because IndexRequest.source(Map) cannot handle
BigInteger values.

This exception is not caught, and triggers a panic that terminates
the connector process.

Modifications
=============
Catch any exception thrown in RequestFactory.newIndexRequest. Treat
it as if Elasticsearch server had rejected the request (write an entry
to the reject log).

Change-Id: Ib68f99352e83d0d273205f6559b88b7b3bf110cd
Reviewed-on: http://review.couchbase.org/106018
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>