CBES-126 Fix rare race condition that only affects integration tests.
Motivation
==========
The tests run multiple connector instances in the same process,
and the instances all compete to configure gauges.

If two connectors try to configure the `cbes.esWaitMs` gauge at the
same time, an exception is thrown with the message:
"cbes.esWaitMs is already used for a different type of metric"

Modifications
=============
Synchronize so only one connector instance can configure the gauges
at a time.

Change-Id: I38fdbcd417ffea61fa4383b169ab6f7f6161adae
Reviewed-on: http://review.couchbase.org/108907
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>