CBES-147 Stuck shutdown hook can prevent/delay JVM termination
Motivation
----------
In some circumstances the connector will attempt to shut itself down
due to a fatal error, but the JVM process does not terminate.

A stuck shutdown hook is the most likely suspect.

Modifications
-------------
Instead of registering shutdown hooks directly, add a new RuntimeHelper
class that manages the hooks and halts the JVM if they do not complete
within a reasonable amount of time.

Shutdown hooks are a "best effort" cleanup mechanism anyway, so the
consequences are minor if we halt before they complete.

Change-Id: I140330e19a494d7c1ff6ef57675800a216604851
Reviewed-on: http://review.couchbase.org/119991
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>