Add Dockerfiles to build Docker images
Motivation
----------
Easily create a Docker image for a particular version of the connector
that is released and distributed.

Modifications
-------------
Add Dockerfile which builds locally using gradle and a multi-stage
Docker build.

Add Dockerfile.light which accepts a build argument for the version of
the connector to download and install in the Docker image.

Updated README.adoc with Docker build instructions, Docker run
instructions, and Kubernetes run instructions.

Results
-------
Anyone can easily create a compatible Docker image to deploy the
connector locally or in Kubernetes. Supports using volume mappings to
mount configuration and secrets, which is compatible with Kubernetes
ConfigMap and Secret volume mounts.

Change-Id: Ideec0ded0a31d3eab31a6ac95d0f4c13e8181815
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/131489
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>