CBES-163 Support deriving index from scope and/or collection name
Motivation
----------
Elasticsearch indexes align quite nicely with the intended use of
Couchbase collections. We want to make it easy for a user to define
a rule that writes documents to an index whose name is the same as
the Couchbase scope and/or collection.

Modifications
-------------
Add a 'matchOnQualifiedKey' flag that controls whether the key name
exposed to the prefix and regex matchers is qualified by scope and
collection. For example, "scope.collection.foo" versus just "foo".

Enabling matching on qualified key lets the type rules act on the
scope and collection.

Add some example rules that use the existing regex capturing group
behavior to derive the index name from the scope and/or collection.

Change-Id: I6e2e892a3ffd52320f6acb1c29de959720424a0e
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/132069
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>