CBES-179 Insecure ES connection with custom port and secureConnection=true
Motivation
----------
Specifying a custom Elasticsearch port causes the client to attempt
an insecure connection unless the hosts string starts with "https://".
This is a problem because the documentation and examples specify
the address as simply "host:port" without a URI prefix.

Modifications
-------------
Fix the parsing, and add regression tests.

Add a sanity check to prohibit explicitly specifying insecure "http"
in URL when secureConnection=true. (Without this check, the previous
behavior was for the `secureConnection` setting to override the user's
specified URI scheme.)

Change-Id: I88251deddb5f63b3330dc60d65fc22444aa67289
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/139499
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>