CBES-185 Support Consul 1.9.1
Motivation
----------
A bug in the Consul client we're using prevents the connector from working
with Consul 1.8.4 or later.

See https://github.com/rickfast/consul-client/issues/431

The client bug has been patched, but we don't know when a new version will
be released. In the mean time, we can apply a workaround.

Modifications
-------------
Test against Consul 1.9.1.

Use a Jackson mixin to prevent the Consul client from parsing the
debug config in the /v1/agent/self response.

Change-Id: I7dbfefc9d66f722f2c88b2bfd7d1a1998c6cd7bb
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/143085
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>