CBES-193 Duplicate eviction does not consider scope/collection
Motivation
----------
Documents with the same base ID were being treated as duplicates,
even if they were destined for different indexes. (They indexes
could be different if their type definition rules matched on the
collection and/or scope.)

Modifications
-------------
Consider the destination index when checking for duplicates.

Log a warning when there is a legitimate document ID conflict
due to misconfiguration.

Future work: See if the misconfiguration can be detected
at startup instead of waiting for a collision to happen.

Change-Id: I3a8e8e9f6d0d1fa85ee588e6ac39fa41c1926448
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/147022
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>