CBES-206 Configuration: Allow specifying int and boolean values as strings
Motivation
----------
Customer wants to use environment variable placeholders for integer
and boolean config properties. Normally this results in a
file that is not valid TOML, but that's fine since the connector
turns it into a valid TOML file by resolving the placeholders.

Example of the invalid TOML:

  numberOfThings = ${NUMBER_OF_THINGS:0}

Customer wants to use other tools to parse the config file,
so they want it to always be valid TOML.

Modification
------------
Allow specifying boolean / integer config properties as strings.
For example, the following are now both valid ways to specify
an integer value:

  numberOfThings = ${NUMBER_OF_THINGS:0}
  numberOfThings = '${NUMBER_OF_THINGS:0}'

This allows users to smuggle the placeholders inside strings
where a TOML parser won't complain about them.

Add a facade over the TOML parser so we can intercept
all calls to getInt() and getBoolean() and handle the case where
the value was specified as a string.

Change-Id: I35530cbe8db025b8423d65741687beebd39df63a
Reviewed-on: http://review.couchbase.org/c/couchbase-elasticsearch-connector/+/152017
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>